# Document/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "GearGo",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/ban-giao/don-thue/{maDonThue}": {
      "get": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ban-giao/don-thue/{maDonThue}": {
      "get": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ban-giao/theo-don/{maDonThue}": {
      "get": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ban-giao/theo-don/{maDonThue}": {
      "get": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ban-giao/nhap/{maDonThue}": {
      "post": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ban-giao/nhap/{maDonThue}": {
      "post": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maDonThue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ban-giao/nhap/{maPhieuBanGiao}": {
      "put": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maPhieuBanGiao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ban-giao/nhap/{maPhieuBanGiao}": {
      "put": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maPhieuBanGiao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ban-giao/{maPhieuBanGiao}": {
      "put": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maPhieuBanGiao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ban-giao/{maPhieuBanGiao}": {
      "put": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maPhieuBanGiao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatPhieuBanGiaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ban-giao/{maPhieuBanGiao}/chot": {
      "post": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maPhieuBanGiao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChotBanGiaoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChotBanGiaoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChotBanGiaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ban-giao/{maPhieuBanGiao}/chot": {
      "post": {
        "tags": [
          "BanGiao"
        ],
        "parameters": [
          {
            "name": "maPhieuBanGiao",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChotBanGiaoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChotBanGiaoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChotBanGiaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/don-thue/{id}": {
      "get": {
        "tags": [
          "DonThue"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/don-thue/{id}/ban-giao": {
      "get": {
        "tags": [
          "DonThue"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/nha-cung-cap": {
      "get": {
        "tags": [
          "NhaCungCap"
        ],
        "parameters": [
          {
            "name": "tuKhoa",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trangThaiHopTac",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trang",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "soMoiTrang",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "NhaCungCap"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaoNhaCungCapRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TaoNhaCungCapRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TaoNhaCungCapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/nha-cung-cap/{id}": {
      "get": {
        "tags": [
          "NhaCungCap"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "NhaCungCap"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatNhaCungCapRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatNhaCungCapRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatNhaCungCapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/nha-cung-cap/{id}/trang-thai-hop-tac": {
      "put": {
        "tags": [
          "NhaCungCap"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoiHopTacRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DoiHopTacRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DoiHopTacRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/nha-cung-cap/{id}/lich-su-nhap": {
      "get": {
        "tags": [
          "NhaCungCap"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "trang",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "soMoiTrang",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapNhatChiTietBanGiaoRequest": {
        "type": "object",
        "properties": {
          "maPhanCong": {
            "type": "integer",
            "format": "int64"
          },
          "tinhTrangTruocThue": {
            "type": "string",
            "nullable": true
          },
          "phuKienThucGiao": {
            "type": "string",
            "nullable": true
          },
          "danhSachAnh": {
            "type": "string",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CapNhatNhaCungCapRequest": {
        "required": [
          "ten"
        ],
        "type": "object",
        "properties": {
          "ten": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "nguoiLienHe": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "soDienThoai": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "email": {
            "maxLength": 255,
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "diaChi": {
            "type": "string",
            "nullable": true
          },
          "maSoThue": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          },
          "lyDoThayDoi": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CapNhatPhieuBanGiaoRequest": {
        "type": "object",
        "properties": {
          "tenNguoiNhanThucTe": {
            "type": "string",
            "nullable": true
          },
          "thoiDiemKhachXacNhan": {
            "type": "string",
            "format": "date-time"
          },
          "bangChungXacNhan": {
            "type": "string",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          },
          "chiTiet": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CapNhatChiTietBanGiaoRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChotBanGiaoRequest": {
        "type": "object",
        "properties": {
          "tenNguoiNhanThucTe": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DoiHopTacRequest": {
        "required": [
          "lyDo",
          "trangThaiHopTac"
        ],
        "type": "object",
        "properties": {
          "trangThaiHopTac": {
            "minLength": 1,
            "pattern": "^(DangHopTac|NgungHopTac)$",
            "type": "string"
          },
          "lyDo": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "TaoNhaCungCapRequest": {
        "required": [
          "maHienThi",
          "ten"
        ],
        "type": "object",
        "properties": {
          "maHienThi": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "ten": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "nguoiLienHe": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "soDienThoai": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "email": {
            "maxLength": 255,
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "diaChi": {
            "type": "string",
            "nullable": true
          },
          "maSoThue": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Nhập trực tiếp chuỗi JWT Token vào ô bên dưới (không cần gõ thêm chữ 'Bearer ')",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}